third_party/WebKit/LayoutTests/fast/events/constructors/media-key-message-event-constructor.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>

description("This tests the constructor for the MediaKeyMessageEvent DOM class.");

// No initializer is passed.
shouldBe("new MediaKeyMessageEvent('eventType').bubbles", "false");
shouldBe("new MediaKeyMessageEvent('eventType').cancelable", "false");
shouldBeEqualToString("new MediaKeyMessageEvent('eventType').messageType", "license-request");
shouldBeNull("new MediaKeyMessageEvent('eventType').message");

// bubbles is passed.
shouldBe("new MediaKeyMessageEvent('eventType', { bubbles: false }).bubbles", "false");
shouldBe("new MediaKeyMessageEvent('eventType', { bubbles: true }).bubbles", "true");

// cancelable is passed.
shouldBe("new MediaKeyMessageEvent('eventType', { cancelable: false }).cancelable", "false");
shouldBe("new MediaKeyMessageEvent('eventType', { cancelable: true }).cancelable", "true");

// message is passed.
var arrayBuffer = new ArrayBuffer(1);
shouldBe("new MediaKeyMessageEvent('eventType', { message: arrayBuffer }).message", "arrayBuffer");

// messageType is passed.
shouldBeEqualToString("new MediaKeyMessageEvent('eventType', { messageType: 'license-request' }).messageType", "license-request");
shouldBeEqualToString("new MediaKeyMessageEvent('eventType', { messageType: 'license-renewal' }).messageType", "license-renewal");
shouldBeEqualToString("new MediaKeyMessageEvent('eventType', { messageType: 'license-release' }).messageType", "license-release");

</script>
</body>
</html>
