third_party/WebKit/LayoutTests/fast/events/constructors/media-key-message-event-constructor-expected.txt

This tests the constructor for the MediaKeyMessageEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new MediaKeyMessageEvent('eventType').bubbles is false
PASS new MediaKeyMessageEvent('eventType').cancelable is false
PASS new MediaKeyMessageEvent('eventType').messageType is "license-request"
PASS new MediaKeyMessageEvent('eventType').message is null
PASS new MediaKeyMessageEvent('eventType', { bubbles: false }).bubbles is false
PASS new MediaKeyMessageEvent('eventType', { bubbles: true }).bubbles is true
PASS new MediaKeyMessageEvent('eventType', { cancelable: false }).cancelable is false
PASS new MediaKeyMessageEvent('eventType', { cancelable: true }).cancelable is true
PASS new MediaKeyMessageEvent('eventType', { message: arrayBuffer }).message is arrayBuffer
PASS new MediaKeyMessageEvent('eventType', { messageType: 'license-request' }).messageType is "license-request"
PASS new MediaKeyMessageEvent('eventType', { messageType: 'license-renewal' }).messageType is "license-renewal"
PASS new MediaKeyMessageEvent('eventType', { messageType: 'license-release' }).messageType is "license-release"
PASS successfullyParsed is true

TEST COMPLETE
